worker/requirements.txt
sqlmodel==0.0.21
psycopg[binary]==3.1.19
boto3==1.34.106
requests==2.32.3
youtube-transcript-api==0.6.2
openai-whisper==20231117
torch==2.2.2
